refactor(sections): extract ParsedHelp emission into emit.rs
Pure motion: flags, subcommands, a flag's enumerated choices and declared
positionals, plus the command-table emitters, with their tests.
